fix(ipc): 格式不一致修正与 stale 代码清理
fix(ipc): format inconsistency fixes and stale code cleanup

- IpcCommandHandler::GetInterfaceId 添加 noexcept，修复 RegisterHandler broken doxygen
- MainProcess/IpcContext.h 删除 GetProxyFactory 返回类型多余的 class keyword
- MainProcess/IIpcContext.h 清理 namespace 前置声明注释
- Handshake.h 删除 5 处重复 interface_id doxygen 注释行
- InterfaceParamSerialization.cpp 移除 3 处不必要的 static_cast<int>
- RemoteObjectRegistry.cpp 两处 "Object not found" 日志补充 encoded_id

～(∠・ω< )⌒★ GLM-5 代表 Dusk 创建了这个提交

+++ das/Core/IPC/src/InterfaceParamSerialization.cpp
@@ -71,7 +71,7 @@ DasResult SerializeInInterfaceParam(
         DAS_CORE_LOG_ERROR(
             "SerializeInInterfaceParam: failed to auto-register local object, "
             "result = {}",
-            static_cast<int>(register_result));
+            register_result);
         return register_result;
     }
     if (out_newly_registered != nullptr)
@@ -127,7 +127,7 @@ DasResult DeserializeInInterfaceParam(
             DAS_CORE_LOG_ERROR(
                 "DeserializeInInterfaceParam: RegisterRemoteObject failed, "
                 "result = {}",
-                static_cast<int>(register_result));
+                register_result);
             return register_result;
         }
         *out_ptr = by_value_proxy.release();
